refactor(bus): split address routing + pin resolution into routing.rs
Move the peripheral-window lookup (find_peripheral_index, rebuild/refresh
ranges, resolve_window, next_window_start, bit_band_translate, ...) and GPIO
pin resolution (resolve_pin_odr/idr, parse_stm32_pin, parse_esp32_gpio_pin)
out of bus/mod.rs into bus/routing.rs. Methods called from sibling modules
(accessors, from_config) made pub(crate).

bus/mod.rs 2038 -> 1795 (3509 -> 1795 across 4 bus PRs). Verified both feature
sets: workspace lib + jit,event-scheduler lib + firmware_survival + register_coverage.
